Migrate from LambdaControls to MidnightControls
