+++ gradle.properties
@@ -8,7 +8,9 @@ loader_version=0.13.3
 fabric_version=0.51.1+1.18.2
 
 # Other Dependencies
-lambdacontrols_version=1.7.1+1.17
+midnightcontrols_version=1.0.0
+midnightlib_version=0.4.0
+spruceui_version=3.3.3+1.18
 myron_version=1.6.3+1.18.1
 arrp_version=0.5.5
 jsonem_version=0.1.2
