+++ src/main/java/io/github/foundationgames/automobility/mixin/midnightcontrols/InputHandlersMixin.java
@@ -1,9 +1,9 @@
-package io.github.foundationgames.automobility.mixin.lambdacontrols;
+package io.github.foundationgames.automobility.mixin.midnightcontrols;
 
-import dev.lambdaurora.lambdacontrols.client.controller.ButtonBinding;
-import dev.lambdaurora.lambdacontrols.client.controller.InputHandlers;
+import eu.midnightdust.midnightcontrols.client.controller.ButtonBinding;
+import eu.midnightdust.midnightcontrols.client.controller.InputHandlers;
 import io.github.foundationgames.automobility.entity.AutomobileEntity;
-import io.github.foundationgames.automobility.util.lambdacontrols.AutomobilityLC;
+import io.github.foundationgames.automobility.util.midnightcontrols.AutomobilityMidnightControls;
 import net.minecraft.client.MinecraftClient;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.asm.mixin.Mixin;
@@ -21,7 +21,7 @@ public class InputHandlersMixin {
     private static void automobility$makeAutomobileInputsWork(@NotNull MinecraftClient client, @NotNull ButtonBinding binding, CallbackInfoReturnable<Boolean> cir) {
         var player = client.player;
         if (!(player == null || !(player.getVehicle() instanceof AutomobileEntity))) {
-            for (ButtonBinding ab : AutomobilityLC.AUTOMOBILITY_BINDINGS) {
+            for (ButtonBinding ab : AutomobilityMidnightControls.AUTOMOBILITY_BINDINGS) {
                 if (Arrays.equals(ab.getButton(), binding.getButton())) cir.setReturnValue(false);
             }
         }
